# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nutrition Clinic API",
    "version": "2.0.0",
    "description": "واجهة REST لنظام إدارة عيادة التغذية. مصمَّمة ليستهلكها موقع الويب وتطبيق الموبايل (Android/iOS) الخاص بالمريض دون أي تعديل. المصادقة: POST /api/auth/login ثم إرسال `Authorization: Bearer <token>` في كل طلب. التوكن صالح 12 ساعة.",
    "contact": {
      "name": "Clinic IT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "تطوير محلي"
    },
    {
      "url": "https://api.clinic.example",
      "description": "الإنتاج"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "تسجيل الدخول",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "username",
                  "password"
                ],
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "{ token, token_type, expires_in, user }"
          },
          "400": {
            "description": "بيانات دخول خاطئة"
          },
          "409": {
            "description": "حساب موقوف"
          }
        }
      }
    },
...
    "/api/patients": {
      "get": {
        "tags": [
          "patients"
        ],
        "summary": "قائمة + بحث سريع",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "اسم أو رقم ملف أو هاتف"
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "active",
                "inactive",
                "archived"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "file_no",
                "name",
                "created"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{ total, items: Patient[] }"
          }
        }
      },
      "post": {
        "tags": [
          "patients"
        ],
        "summary": "إنشاء مريض (رقم الملف تلقائي)",
        "responses": {
          "201": {
            "description": "Patient"
          },
          "400": {
            "description": "تحقق المدخلات"
          }
        }
      }
    },
...
    "/api/diet-plans": {
      "get": {
        "tags": [
          "plans"
        ],
        "summary": "البرامج الغذائية",
        "parameters": [
          {
            "name": "patient_id",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{ items }"
          }
        }
      },
      "post": {
        "tags": [
          "plans"
        ],
        "summary": "إنشاء برنامج مع الوجبات (المجاميع تُحسب آلياً)",
        "responses": {
          "201": {
            "description": "DietPlan"
          }
        }
      }
    },
...
    "/api/appointments": {
      "get": {
        "tags": [
          "appointments"
        ],
        "summary": "قائمة مواعيد مع فلاتر",
        "parameters": [
          {
            "name": "patient_id",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "scheduled",
                "confirmed",
                "done",
                "cancelled",
                "no_show"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{ items }"
          }
        }
      },
      "post": {
        "tags": [
          "appointments"
        ],
        "summary": "تسجيل موعد (يرفض التداخل 409)",
        "responses": {
          "201": {
            "description": "Appointment"
          },
          "409": {
            "description": "الوقت محجوز"
          }
        }
      }
    },
...
    "/api/payments": {
      "get": {
        "tags": [
          "payments"
        ],
        "summary": "المدفوعات مع الإجمالي",
        "parameters": [
          {
            "name": "patient_id",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{ items, total, currency }"
          }
        }
      },
      "post": {
        "tags": [
          "payments"
        ],
        "summary": "تسجيل دفعة",
        "responses": {
          "201": {
            "description": "Payment"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  }
}